+++ rpgRogueBeast/beastmaster/BullPet.java
@@ -71,7 +71,7 @@ public class BullPet extends BMPetImpl {
         return (4 + MathHelper.floor_double((((double) getLevel()) * /*Fix for math errors*/1.0D) / 18.18D));
     }
     @Override
-    public int getMaxHealth() {
+    public float getMaxHealth() {
         //200 HP at level 200
         return 30 + MathHelper.floor_float(((float) getLevel()) / 1.538F);
     }
